refactor: moved tests to seperate folder
