+++ package.json
@@ -4,10 +4,11 @@
   "description": "A GitHub Action to suggest and enforce Conventional Commits.",
   "main": "dist/index.js",
   "scripts": {
-    "build": "tsc",
+    "build": "ncc build src/index.ts -o dist",
+    "typecheck": "tsc --noEmit -p tsconfig.test.json",
     "lint": "eslint . --ext .ts",
-    "typecheck": "tsc --noEmit",
-    "test": "jest"
+    "test": "jest",
+    "prepare": "npm run build"
   },
   "keywords": [
     "github-action",
@@ -26,14 +27,15 @@
     "@semantic-release/git": "^10.0.1",
     "@semantic-release/github": "^11.0.3",
     "@types/jest": "^29.5.14",
-    "@types/node": "^20.0.0",
+    "@types/node": "^20.19.7",
     "@typescript-eslint/eslint-plugin": "^8.35.1",
     "@typescript-eslint/parser": "^8.35.1",
+    "@vercel/ncc": "^0.38.3",
     "conventional-changelog-conventionalcommits": "^9.1.0",
     "eslint": "^8.57.1",
-    "jest": "^29.6.0",
+    "jest": "^29.7.0",
     "semantic-release": "^24.2.6",
     "ts-jest": "^29.1.1",
-    "typescript": "^5.4.0"
+    "typescript": "^5.8.3"
   }
 }
