+++ tsconfig.json
@@ -13,7 +13,6 @@
     "typeRoots": ["./src/types", "./node_modules/@types"]
   },
   "include": [
-    "src",
-    "node_modules/@types/jest/index.d.ts"
+    "src"
   ]
 }
